add watchmaker test
